feat(relay): liftOptional as the Absent-action/environment scope host
Redefine liftOptional to consume a state-only Relay.Scope — Absent action +
Absent environment + an affine Writes state lane — so the compiler enforces the
everything-absent-but-state shape. Preserves the eager-guard semantics (a
complete no-op while the focus is nil: no mutation, no effect, no supervise —
stricter than a plain affine state lift, which would forward the nil and still
run). The key-path form is kept as delegating sugar. Behavior + Middleware only
(a store can never be absent).
